+++ oqstests/sig_test.go
@@ -42,7 +42,28 @@ func testSigCorrectness(sigName string, msg []byte, threading bool, t *testing.T
 	isValid, _ := verifier.Verify(msg, signature, pubKey)
 	if !isValid {
 		// t.Errorf is thread-safe
-		t.Errorf(sigName + ": signature verification failed")
+		t.Errorf("%s: signature verification failed", sigName)
+	}
+}
+
+// testSigCorrectness tests a specific signature with context string.
+func testSigCorrectnessWithCtxStr(sigName string, msg []byte, threading bool, t *testing.T) {
+	log.Println("Correctness - ", sigName) // thread-safe
+	if threading == true {
+		defer wgSigCorrectness.Done()
+	}
+	var signer, verifier oqs.Signature
+	defer signer.Clean()
+	defer verifier.Clean()
+	// Ignore potential errors everywhere
+	_ = signer.Init(sigName, nil)
+	_ = verifier.Init(sigName, nil)
+	pubKey, _ := signer.GenerateKeyPair()
+	signature, _ := signer.Sign(msg)
+	isValid, _ := verifier.Verify(msg, signature, pubKey)
+	if !isValid {
+		// t.Errorf is thread-safe
+		t.Errorf("%s: signature verification failed", sigName)
 	}
 }
 
@@ -64,7 +85,7 @@ func testSigWrongSignature(sigName string, msg []byte, threading bool, t *testin
 	isValid, _ := verifier.Verify(msg, wrongSignature, pubKey)
 	if isValid {
 		// t.Errorf is thread-safe
-		t.Errorf(sigName + ": signature verification should have failed")
+		t.Errorf("%s: signature verification should have failed", sigName)
 	}
 }
 
@@ -86,7 +107,7 @@ func testSigWrongPublicKey(sigName string, msg []byte, threading bool, t *testin
 	isValid, _ := verifier.Verify(msg, signature, wrongPubKey)
 	if isValid {
 		// t.Errorf is thread-safe
-		t.Errorf(sigName + ": signature verification should have failed")
+		t.Errorf("%s: signature verification should have failed", sigName)
 	}
 }